Split remote PR command builders
