+++ docs/DECISIONS.md
@@ -385,7 +385,7 @@
 - Git hunk patch execution lives in `GitPatchToolExecutor.swift`, and git tool errors live in `GitToolError.swift`. Hunk staging/restoring, patch temp-file handling, and patch-path metadata parsing should stay beside the patch executor; local and SSH Remote hunk planning should reuse `GitPatchToolExecutor.mismatchedPatchPath` instead of adding patch parsers to facades or remote planners.
 - Local git execution lives in `GitLocalToolExecutor.swift`. Status, diff, stage, restore, commit, push, path validation, and current-branch lookup should stay beside the local executor; `GitToolExecutor` remains only the compatibility facade over local, patch, worktree, and GitHub PR executors.
 - GitHub PR input validation and output parsing live outside PR command execution. `GitHubPullRequestInputValidator.swift` owns selector, reviewer, label, review-action, and merge-method normalization for both local and SSH Remote PR planning, while `GitHubPullRequestOutputParser.swift` owns URL artifact extraction. `GitHubPullRequestToolExecutor.swift` should stay focused on `gh pr` command construction and execution.
-- SSH Remote GitHub PR command construction lives in `WorkspaceRemoteGitHubPullRequestCommandBuilder.swift`. `WorkspaceRemoteGitToolRequestPlanner.swift` should route to that builder and keep generic git, hunk, worktree, and artifact intent planning separate so local PR validation, SSH Remote command assembly, and future QuillCloud remote transports do not drift.
+- SSH Remote GitHub PR command construction has a stable facade plus focused builders. `WorkspaceRemoteGitHubPullRequestCommandBuilder.swift` owns tool-family routing and URL artifact intent only; primary PR commands live in `WorkspaceRemoteGitHubPullRequestPrimaryCommandBuilder.swift`, reviewer/label/comment/review commands live in `WorkspaceRemoteGitHubPullRequestCollaborationCommandBuilder.swift`, review-thread API commands live in `WorkspaceRemoteGitHubPullRequestReviewThreadCommandBuilder.swift`, and shared PR metadata/quoting snippets live in `WorkspaceRemoteGitHubPullRequestCommandSupport.swift`. `WorkspaceRemoteGitToolRequestPlanner.swift` should route to the facade and keep generic git, hunk, worktree, and artifact intent planning separate so local PR validation, SSH Remote command assembly, and future QuillCloud remote transports do not drift.
 - SSH Remote git worktree command construction lives in `WorkspaceRemoteGitWorktreeCommandBuilder.swift`. Worktree list/create/open/remove command assembly, sibling path normalization, registered-worktree verification, and worktree-artifact reporting should stay in that builder while `WorkspaceRemoteGitToolRequestPlanner.swift` remains a generic router for remote git command families.
 - SSH Remote git hunk command construction lives in `WorkspaceRemoteGitHunkCommandBuilder.swift`. Stage/restore hunk patch validation, base64 patch transport, temporary-file cleanup, and `git apply` check/apply command assembly should stay in that builder while `WorkspaceRemoteGitToolRequestPlanner.swift` remains a generic router for remote git command families.
 - SSH Remote git push command construction lives in `WorkspaceRemoteGitPushCommandBuilder.swift`. Explicit branch pushes, current-branch detection, current-branch safety guards, upstream flags, and remote/branch validation should stay in that builder while `WorkspaceRemoteGitToolRequestPlanner.swift` remains a generic router for remote git command families.
@@ -462,7 +462,7 @@
 - Native and rendered command registries must stay in parity. App tests parse the rendered harness routing registry and compare it against a rich native `WorkspaceCommandSurfaceBuilder` command set plus every `WorkspacePullRequestCommandCatalog` descriptor, so adding a Swift command now requires adding the corresponding harness static ID or dynamic prefix before Playwright can pass.
 - Local git, git hunk, git worktree, and git router coverage lives in focused QuillCodeTools suites, not the mixed `ToolTests.swift` catch-all. `GitLocalToolExecutorTests`, `GitPatchToolExecutorTests`, `GitWorktreeToolExecutorTests`, and `GitToolRouterTests` share fixtures from `ToolTestSupport.swift` so git behavior can evolve without turning one broad file into a failure triage bottleneck.
 - `ToolTests.swift` is retired. File primitives, generic apply-patch primitives, shell router boundary checks, shell executor behavior, git tools, GitHub PR tools, and MCP probing each live in named focused suites, with `ParityToolGateTests` guarding the retirement of the broad catch-all.
-- Inline GitHub pull request review comments are a first-class tool (`host.git.pr.review_comment`), not a raw shell workaround. Local execution resolves PR number/head commit with `gh pr view --json`, resolves the repository with `gh repo view --json`, and posts through `gh api`; SSH Remote execution uses the same validation but builds a quoted remote command with only resolved metadata variables expanded. Keep selector/path/line/side/body validation in `GitHubPullRequestInputValidator`, local API orchestration in `GitHubPullRequestToolExecutor`, and SSH command construction in `WorkspaceRemoteGitHubPullRequestCommandBuilder`.
+- Inline GitHub pull request review comments are a first-class tool (`host.git.pr.review_comment`), not a raw shell workaround. Local execution resolves PR number/head commit with `gh pr view --json`, resolves the repository with `gh repo view --json`, and posts through `gh api`; SSH Remote execution uses the same validation but builds a quoted remote command with only resolved metadata variables expanded. Keep selector/path/line/side/body validation in `GitHubPullRequestInputValidator`, local API orchestration in `GitHubPullRequestToolExecutor`, and SSH review-thread command construction in `WorkspaceRemoteGitHubPullRequestReviewThreadCommandBuilder`.
 - Static Auto safety argument-word fallback is read-only only. If a user word appears in a tool's JSON arguments, that can mark intent for read-only discovery, but append/destructive tools must match an explicit tool-specific intent rule such as shell run, file write, apply patch, git push, PR, or MCP. This prevents incidental words like `origin`, `disk`, or file names from approving state-changing actions.
 - Static PR intent treats `PR` as a token, not just long-form `pull request`, but PR-specific rules must consider every matching verb before approving a tool. A request like `reply to review comment` should approve the reply tool even though it also contains `review`, while a request like `show PR 42` must not approve push or merge through the generic PR token.
 - Static Auto intent matching treats nearby negation as a clarification boundary per phrase occurrence. Phrases like `do not run`, `don't apply this patch`, `do not push`, and `don't remember this` must not Auto-approve those tool families even if the underlying action word appears in the request, while `do not run whoami; run hostname` can still approve the affirmed command. The static layer does not try to infer an alternate command from negated text; it asks for review/clarification when only negated intent is present.
